pyproject.toml
[project]
name = "whisper-transcriber"
version = "0.1.0"
description = "Local audio transcription using OpenAI Whisper"
requires-python = ">=3.10, <3.14"
dependencies = [
    "openai-whisper",
    "torch==2.7.1",
    "ffmpeg-python==0.2.0",
    "rich",
    "questionary",
    "google-genai",
]

[project.scripts]
transcriber = "src.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
